Modular Architecture Setup
