+++ examples/floatingUI/floatingUI.css
@@ -1,0 +1,192 @@
+* {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+  }
+  body {
+    min-height: 100vh;
+    background-color: #1111;
+    overflow: hidden;
+  }
+  nav {
+    position: fixed;
+    top: 20px;
+    right: 0;
+    right: calc(100% - 100px);
+    width: 80px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    cursor: grab;
+    left: 150px;
+  }
+  nav .nav-content {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+  }
+  .nav-content .toggle-btn,
+  .nav-content span a {
+    height: 60px;
+    width: 60px;
+    background-color: #fff;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    border-radius: 50%;
+    box-shadow: 0 0 20px rgba(0, 0, 0, 0.2);
+  }
+  .nav-content .toggle-btn {
+    font-size: 35px;
+    color: #0e2431;
+    z-index: 100;
+    cursor: pointer;
+    transition: all 0.6s ease;
+  }
+  .nav-content .myspan {
+    position: absolute;
+    transition: all 0.6s ease;
+    opacity: 0;
+  }
+  nav.open .nav-content span {
+    opacity: 1;
+    transform: translateX(calc(var(--i) * 80px));
+  }
+  .nav-content span a {
+    text-decoration: none;
+  }
+  .nav-content span a i {
+    font-size: 24px;
+    color: #0e2431;
+    opacity: 0.8;
+    transition: 0.2s;
+  }
+  .nav-content span a:hover i {
+    opacity: 1;
+  }
+  .search-box {
+    position: absolute;
+    top: 130%;
+    width: 200px;
+    height: 60px;
+    background-color: white;
+    border-radius: 10px;
+    display: none;
+    justify-content: center;
+    align-items: center;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+  }
+  .search-box.show {
+    display: flex;
+  }
+  input[type="text"] {
+    width: 80%;
+    padding: 10px;
+    font-size: 16px;
+    border: 2px solid #ccc;
+    border-radius: 5px;
+  }
+  .rect-box {
+    position: absolute;
+    top: 130%;
+    width: 300px;
+    height: 150px;
+    background-color: rgb(250, 250, 255);
+    border-radius: 10px;
+    display: none;
+    justify-content: center;
+    align-items: center;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+    flex-direction: column;
+  }
+  .rect-box.show {
+    display: flex;
+  }
+  .rect-box .rect-btn {
+    padding: 8px 16px;
+    margin: 4px;
+    box-sizing: border-box;
+    font-size: 16px;
+    border: none;
+    border-radius: 5px;
+    background-color: #0e2431;
+    color: white;
+    cursor: pointer;
+    transition: background-color 0.3s ease;
+  }
+  .rect-box .rect-btn:hover {
+    background-color: #1a4567;
+  }
+  .button-row {
+    display: flex;
+    width: 100%;
+    justify-content: center;
+  }
+  .rect-box.ML,
+  .rect-box.LLM,
+  .rect-box.Filters,
+  .rect-box.DataViz {
+    display: none;
+  }
+  .rect-box.LLM.show,
+  .rect-box.Filters.show,
+  .rect-box.DataViz.show {
+    display: flex;
+  }
+  .back-arrow {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    cursor: pointer;
+    font-size: 24px;
+    margin-top: 5px;
+    visibility: visibile;
+    color: #333;
+  }
+  .edit-box {
+    position: absolute;
+    top: 130%;
+    width: 700px;
+    height: 200px;
+    padding: 10px;
+    background-color: white;
+    border-radius: 10px;
+    display: none;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+  }
+  #editor {
+    position: relative;
+    height: 100%;
+    width: 100%;
+    border: 1px solid lightgray;
+    margin: auto;
+  }
+  #output {
+    width: 100%;
+    height: 100%;
+    border: 1px solid #ccc;
+    padding: 5px;
+    overflow: auto;
+    color: #333;
+  }
+  #suggestionsContainer {
+    background-color: white;
+    border: 1px solid #ccc;
+    border-top: none;
+    position: absolute;
+    width: 100%;
+    box-shadow: 0px 4px 5px -2px rgba(0, 0, 0, 0.25);
+    z-index: 5;
+  }
+  .suggestion-item {
+    padding: 10px;
+    border-bottom: 1px solid #eee;
+    cursor: pointer;
+  }
+  .suggestion-item:hover {
+    background-color: #f0f0f0;
+  }
+
